fix: fixed all tests and updated folder structure for atomic components
