+++ package.json
@@ -52,6 +52,7 @@
     "@storybook/addon-essentials": "^6.2.9",
     "@storybook/addon-links": "^6.2.9",
     "@storybook/vue": "^6.2.9",
+    "@testing-library/vue": "^5.8.3",
     "@types/jest": "^24.0.19",
     "@types/node": "^18.7.21",
     "@typescript-eslint/eslint-plugin": "^4.18.0",
