--- public/figmaThemes/themes.json
@@ -1,1 +1,0 @@
-{"themes":["base","nano","amp","nikola"]}
