--- src/components v2/Forms/Atoms/Checkbox/theme.txt
@@ -1,39 +1,0 @@
---checkbox-default-icon-name: square;
-    --checkbox-default-icon-font-weight: Light;
-    --checkbox-default-icon-color: #7f7f7f;
-    --checkbox-default-label-color: #000000;
-    --checkbox-default-legacy: icon-fg-color: #ffffff;
-    --checkbox-hover-icon-color: #7f7f7f;
-    --checkbox-hover-label-color: #000000;
-    --checkbox-focus-icon-color: #7f7f7f;
-    --checkbox-focus-label-color: #000000;
-    --checkbox-focus-border-color: rgba(#6475f7, 25%);
-    --checkbox-checked-icon-color: #3d53f5;
-    --checkbox-checked-icon-font-weight: Solid;
-    --checkbox-checked-icon-name: square-check;
-    --checkbox-checked-label-color: #000000;
-    --checkbox-disabled-icon-color: #bdbdbd;
-    --checkbox-disabled-label-color: #bdbdbd;
-    --checkbox-indeterminate-icon-name: square-minus;
-    --checkbox-indeterminate-icon-font-weight: Solid;
-    --checkbox-md-label-typography-fontFamily: Inter;
-    --checkbox-md-label-typography-fontWeight: Regular;
-    --checkbox-md-label-typography-fontSize: 16;
-    --checkbox-md-label-typography-lineHeight: 150%;
-    --checkbox-md-label-typography-letterSpacing: 0%;
-    --checkbox-md-label-typography-paragraphSpacing: 0;
-    --checkbox-md-label-typography-textDecoration: none;
-    --checkbox-md-label-typography-textCase: none;
-    --checkbox-md-border-radius: 4;
-    --checkbox-sm-label-typography-fontFamily: Inter;
-    --checkbox-sm-label-typography-fontWeight: Regular;
-    --checkbox-sm-label-typography-fontSize: 14;
-    --checkbox-sm-label-typography-lineHeight: 150%;
-    --checkbox-sm-label-typography-letterSpacing: 0%;
-    --checkbox-sm-label-typography-paragraphSpacing: 0;
-    --checkbox-sm-label-typography-textDecoration: none;
-    --checkbox-sm-label-typography-textCase: none;
-    --checkbox-sm-border-radius: 2;
-    --checkbox-focus-checked-icon-color: #3d53f5;
-    --checkbox-focus-checked-label-color: #000000;
-    --checkbox-focus-checked-border-color: rgba(#6475f7, 25%);
