+++ src/components v2/Molecules/Forms/InputCheckbox/index.ts
@@ -1,0 +1,3 @@
+import P2InputCheckbox from './InputCheckbox.vue';
+
+export default P2InputCheckbox;
