+++ src/components v2/Molecules/Forms/InputCheckboxGroup/index.ts
@@ -1,0 +1,3 @@
+import P2InputCheckboxGroup from './InputCheckboxGroup.vue';
+
+export default P2InputCheckboxGroup;
